% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternativeSplicingEvents.R
\docType{package}
\name{alternativeSplicingEvents}
\alias{alternativeSplicingEvents}
\title{Alternative splicing annotation for AnnotationHub}
\source{
\url{https://github.com/vastgroup/vast-tools#vastdb-libraries}
}
\description{
Alternative splicing annotation for multiple species. These annotations are
based on the
\href{https://github.com/vastgroup/vast-tools#vastdb-libraries}{VastDB} files used
by the alternative splicing quantification program
\href{https://github.com/vastgroup/vast-tools}{VAST-TOOLS}.
}
\details{
Annotation data is available for the following species and assemblies:
\itemize{
\item Homo sapiens (hg19, Hsa)
\item Homo sapiens (hg38 Hs2)
\item Mus musculus (mm9, Mmu)
\item Mus musculus (mm10, Mm2)
\item Bos taurus (bosTau6, Bta)
\item Gallus gallus (galGal3, Gg3)
\item Gallus gallus (galGal4, Gg4)
\item Xenopus tropicalis (xenTro3, Xt1)
\item Danio rerio (danRer10, Dre)
\item Branchiostoma lanceolatum (braLan2, Bl1)
\item Strongylocentrotus purpuratus (strPur4, Spu)
\item Drosophila melanogaster (dm6, Dme)
\item Strigamia maritima (strMar1, Sma)
\item Caenorhabditis elegans (ce11, Cel)
\item Schmidtea mediterranea (schMed31, Sme)
\item Nematostella vectensis (nemVec1, Nve)
\item Arabidopsis thaliana (araTha10, Ath)
}

The available alternative splicing event types include:
\itemize{
\item Skipped exon
\item Retained intron
\item Alternative 3' splice site
\item Alternative 5' splice site
}

Splicing events not assigned to any genes are marked as \code{Hypothetical}.

For more details, check the \code{scripts/make-data.R} file.
}
\references{
\strong{VAST-TOOLS:} Irimia M, Weatheritt RJ, Ellis J, \emph{et al.} \href{http://doi.org/10.1016/j.cell.2014.11.035}{A highly conserved program of neuronal microexons is misregulated in autistic brains.}
Cell. 2014;159(7):1511-1523.
}
